requirements.txt
streamlit>=1.24.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0,<4.12.0
reportlab>=4.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.25.0,<2.0.0
requests>=2.31.0
plotly>=5.24.0
kaleido>=0.2.1
polars[rtcompat]>=1.0.0
